angel_md_redis/requirements.txt
smartapi-python
pyotp
pandas
requests
redis
pyarrow
